+++ server/src/api/controllers/interfaces/IAdminRequest.ts
@@ -1,0 +1,9 @@
+import IReqParams from "./IReqParams";
+import IAdminReqBody from "./IAdminReqBody";
+
+interface IAdminRequest {
+  params: IReqParams;
+  body: IAdminReqBody;
+}
+
+export default IAdminRequest;