+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrading notes for brands. It's a team effort to make them as straightforward as possible.
 
+## [1.4.0] - 2022-06-30
+
+### Added
+
+- Added login system for admin 
+- Require admin name and password for login (.env)
+- Using JWT token
+
 ## [1.3.1] - 2022-06-30
 
 ### Fixed